requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.6.0=pypi_0
blis=0.7.11=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.12.12=haa95532_0
catalogue=2.0.10=pypi_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
colorama=0.4.6=pypi_0
confection=0.1.4=pypi_0
contourpy=1.2.0=pypi_0
cryptography=42.0.2=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
expat=2.5.0=hd77b12b_0
fonttools=4.47.2=pypi_0
idna=3.6=pypi_0
jinja2=3.1.3=pypi_0
kiwisolver=1.4.5=pypi_0
langcodes=3.3.0=pypi_0
libffi=3.4.4=hd77b12b_0
markupsafe=2.1.4=pypi_0
matplotlib=3.8.2=pypi_0
murmurhash=1.0.10=pypi_0
networkx=3.2.1=pypi_0
numpy=1.26.3=pypi_0
openssl=3.0.12=h2bbff1b_0
packaging=23.2=pypi_0
pandas=2.2.0=pypi_0
pdfminer-six=20231228=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py312haa95532_0
preshed=3.0.9=pypi_0
pycparser=2.21=pypi_0
pycryptodome=3.20.0=pypi_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pyparsing=3.1.1=pypi_0
python=3.12.1=h1d929f7_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
requests=2.31.0=pypi_0
setuptools=68.2.2=py312haa95532_0
six=1.16.0=pypi_0
smart-open=6.4.0=pypi_0
spacy=3.7.2=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.41.2=h2bbff1b_0
srsly=2.4.8=pypi_0
thinc=8.2.2=pypi_0
tk=8.6.12=h2bbff1b_0
tqdm=4.66.1=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.4=pypi_0
urllib3=2.1.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wasabi=1.1.2=pypi_0
weasel=0.3.4=pypi_0
wheel=0.41.2=py312haa95532_0
xz=5.4.5=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
